+++ src/inputs/spectrum.rs
@@ -1,19 +1,59 @@
 use std::sync::Arc;
 
-use cpal::Stream;
+use cpal::{
+    Stream,
+    traits::{DeviceTrait, HostTrait, StreamTrait},
+};
 use ringbuf::{
     HeapRb,
-    traits::{Consumer, Observer, RingBuffer},
+    traits::{Consumer, Observer, Producer, RingBuffer, Split},
     wrap::caching::Caching,
 };
 use spectrum_analyzer::scaling::*;
 use spectrum_analyzer::windows::hann_window;
 use spectrum_analyzer::{FrequencyLimit, samples_fft_to_spectrum};
 
-use crate::{
-    audio_stream::{self, NUM_SAMPLES},
-    config::SpectrumConfig,
-};
+use crate::config::SpectrumConfig;
+
+const NUM_SAMPLES: usize = 2048;
+
+// Audio stream setup function
+fn setup_audio_stream() -> (Stream, Caching<Arc<HeapRb<f32>>, false, true>) {
+    let host = cpal::default_host();
+    let input_device = host
+        .default_input_device()
+        .expect("failed to find input device");
+    let mut supported_config_range = input_device
+        .supported_input_configs()
+        .expect("error while querying configs");
+    let supported_config = supported_config_range
+        .next()
+        .expect("no supported config")
+        .with_max_sample_rate();
+    let config = supported_config.into();
+
+    let ring_buffer = HeapRb::<f32>::new(NUM_SAMPLES * 2);
+    let (mut prod, cons) = ring_buffer.split();
+    for _ in 0..NUM_SAMPLES {
+        prod.try_push(0.0).unwrap();
+    }
+
+    let stream = input_device
+        .build_input_stream(
+            &config,
+            move |data: &[f32], _: &cpal::InputCallbackInfo| {
+                for d in data {
+                    if let Ok(()) = prod.try_push(*d) {}
+                }
+            },
+            move |_err| {},
+            None,
+        )
+        .unwrap();
+
+    stream.play().unwrap();
+    (stream, cons)
+}
 
 #[derive(Debug, Clone, Copy)]
 pub(crate) struct SpectrumFrameData {
@@ -132,7 +172,7 @@ impl SpectrumInputManager {
             mapped_at_creation: false,
         });
 
-        let (stream, consumer) = audio_stream::setup_audio_stream();
+        let (stream, consumer) = setup_audio_stream();
 
         Self {
             history_data,
